+++ NEWS.md
@@ -1,5 +1,7 @@
 # mlr3fselect 0.7.2.9000
 
+* refactor: `FSelectorRFE` throws an error if the learner does not support the `$importance()` method.
+
 # mlr3fselect 0.7.2
 
 * docs: Re-generate rd files with valid html.